survey-backend/apps/commons/models/user.py
from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin, BaseUserManager
from django.db import models
from model_utils.models import TimeStampedModel


class UserManager(BaseUserManager):
    def create_user(self, email, password=None, **extra_fields):
        if not email:
            raise ValueError('The Email field must be set')
        email = self.normalize_email(email)
        user = self.model(email=email, **extra_fields)
        user.set_password(password)
        user.save(using=self._db)
        return user

    def create_superuser(self, email, password=None, **extra_fields):
        extra_fields.setdefault('is_staff', True)
        extra_fields.setdefault('is_superuser', True)
        extra_fields.setdefault('firstname', 'Admin')
        extra_fields.setdefault('lastname', 'User')

        if extra_fields.get('is_staff') is not True:
            raise ValueError('Superuser must have is_staff=True.')
        if extra_fields.get('is_superuser') is not True:
            raise ValueError('Superuser must have is_superuser=True.')

        return self.create_user(email, password, **extra_fields)


class User(AbstractBaseUser, PermissionsMixin, TimeStampedModel):
    email = models.EmailField(max_length=254, unique=True, db_index=True)
    firstname = models.CharField(max_length=150, default='')
    lastname = models.CharField(max_length=150, default='')
    is_staff = models.BooleanField(default=False)
    is_superuser = models.BooleanField(default=False)
    is_active = models.BooleanField(default=True)
    last_login = models.DateTimeField(null=True, blank=True)

    USERNAME_FIELD = "email"
    EMAIL_FIELD = "email"
    REQUIRED_FIELDS = []

    objects = UserManager()

    class Meta:
        verbose_name = "User"
        verbose_name_plural = "Users"
        db_table = "users"

    def __repr__(self):
        return f"User({self.pk}, {self.email})"
    
    def __str__(self):
        if self.firstname and self.lastname:
            return f"{self.firstname} {self.lastname} ({self.email})"
        return self.email

    def get_absolute_url(self):
        return f"/users/{self.pk}/"
    
    @property
    def full_name(self):
        """Returns the user's full name."""
        return f"{self.firstname} {self.lastname}".strip() or self.email
