survey-backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2
django-model-utils==5.0.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
drf-spectacular==0.27.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3
pillow==11.2.1
setuptools>=65.0.0
wheel>=0.40.0

# date utilities
python-dateutil==2.8.2

# Data processing and file handling
pandas==2.2.3
openpyxl==3.1.2
python-magic==0.4.27
numpy==1.26.4
scikit-learn==1.5.2
